+++ packages/frontend/pages/settings/components/customize/CustomizeAppearance.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Center, Icon, Radio, VStack } from '@chakra-ui/react';
+import { Box, Center, Icon, Radio, Stack } from '@chakra-ui/react';
 import { Formik, Form } from 'formik';
 import MoonIcon from '@heroicons/react/solid/MoonIcon';
 import SunIcon from '@heroicons/react/solid/SunIcon';
@@ -8,15 +8,19 @@ import {
   AvatarStyles,
   CollectionLayout,
   CollectionLayouts,
+  defaultArticleFontFamily,
+  defaultArticleFontSize,
+  defaultArticleMonoFontFamily,
   defaultAvatarStyle,
   defaultCollectionLayout,
 } from '@orpington-news/shared';
 import { RadioGroupField, AutoSave } from '@components/forms';
 import { CollectionLayoutName } from '@components/collection/types';
 import { RadioButton } from './RadioButton';
+import { ArticleSettings } from './ArticleSettings';
+import { CustomizeAppearanceData, ThemeOption, themeOptions } from './types';
+import { ArticlePreview } from './ArticlePreview';
 
-const themeOptions = ['dark', 'light'] as const;
-export type ThemeOption = typeof themeOptions[number];
 const getThemeIcon = (theme: ThemeOption): React.FC => {
   switch (theme) {
     case 'dark':
@@ -26,16 +30,13 @@ const getThemeIcon = (theme: ThemeOption): React.FC => {
   }
 };
 
-export interface CustomizeAppearanceData {
-  theme: ThemeOption;
-  defaultCollectionLayout: CollectionLayout;
-  avatarStyle: AvatarStyle;
-}
-
 const initialValues: CustomizeAppearanceData = {
   theme: 'dark',
   defaultCollectionLayout: defaultCollectionLayout,
   avatarStyle: defaultAvatarStyle,
+  articleFontFamily: defaultArticleFontFamily,
+  articleMonoFontFamily: defaultArticleMonoFontFamily,
+  articleFontSize: defaultArticleFontSize,
 };
 
 export interface CustomizeAppearanceProps {
@@ -57,53 +58,72 @@ export const CustomizeAppearance: React.FC<CustomizeAppearanceProps> = (
     <Formik initialValues={currentData ?? initialValues} onSubmit={onChange}>
       {({ values }) => (
         <Form noValidate>
-          <VStack spacing={4} w="full">
-            <RadioGroupField label="Theme" name="theme">
-              {({ getRadioProps }) =>
-                themeOptions.map((value) => {
-                  const radio = getRadioProps({ value });
-                  return (
-                    <RadioButton key={value} {...radio}>
-                      <Center p={3}>
-                        <Icon as={getThemeIcon(value)} boxSize={6} />
-                      </Center>
-                    </RadioButton>
-                  );
-                })
-              }
-            </RadioGroupField>
-
-            <RadioGroupField
-              label="Default collection layout"
-              name="defaultCollectionLayout"
+          <Stack
+            spacing={{ base: 4, lg: 8 }}
+            w="full"
+            align="flex-start"
+            direction={{ base: 'column', lg: 'row' }}
+          >
+            <Stack
+              spacing={4}
+              align="flex-start"
+              direction="column"
+              flexShrink={0}
+              flexGrow={1}
             >
-              {({ getRadioProps }) =>
-                CollectionLayouts.map((value: CollectionLayout) => {
-                  const radio = getRadioProps({ value });
-                  return (
-                    <Radio key={value} {...radio}>
-                      {CollectionLayoutName[value]}
-                    </Radio>
-                  );
-                })
-              }
-            </RadioGroupField>
+              <RadioGroupField label="Theme" name="theme">
+                {({ getRadioProps }) =>
+                  themeOptions.map((value) => {
+                    const radio = getRadioProps({ value });
+                    return (
+                      <RadioButton key={value} {...radio}>
+                        <Center p={3}>
+                          <Icon as={getThemeIcon(value)} boxSize={6} />
+                        </Center>
+                      </RadioButton>
+                    );
+                  })
+                }
+              </RadioGroupField>
 
-            <RadioGroupField label="Avatar style" name="avatarStyle">
-              {({ getRadioProps }) =>
-                AvatarStyles.map((value: AvatarStyle) => {
-                  const radio = getRadioProps({ value });
-                  return (
-                    <Radio key={value} {...radio}>
-                      {AvatarStylesNames[value]}
-                    </Radio>
-                  );
-                })
-              }
-            </RadioGroupField>
+              <RadioGroupField
+                label="Default collection layout"
+                name="defaultCollectionLayout"
+              >
+                {({ getRadioProps }) =>
+                  CollectionLayouts.map((value: CollectionLayout) => {
+                    const radio = getRadioProps({ value });
+                    return (
+                      <Radio key={value} {...radio}>
+                        {CollectionLayoutName[value]}
+                      </Radio>
+                    );
+                  })
+                }
+              </RadioGroupField>
+
+              <RadioGroupField label="Avatar style" name="avatarStyle">
+                {({ getRadioProps }) =>
+                  AvatarStyles.map((value: AvatarStyle) => {
+                    const radio = getRadioProps({ value });
+                    return (
+                      <Radio key={value} {...radio}>
+                        {AvatarStylesNames[value]}
+                      </Radio>
+                    );
+                  })
+                }
+              </RadioGroupField>
+
+              <ArticleSettings />
+            </Stack>
+
+            <Box w="full" display={{ base: 'none', md: 'block' }}>
+              <ArticlePreview />
+            </Box>
 
             <AutoSave />
-          </VStack>
+          </Stack>
         </Form>
       )}
     </Formik>