+++ packages/frontend/theme/fonts.ts
@@ -5,8 +5,8 @@ export const fontFaces = `
   font-style: normal;
   font-weight: 400;
   src: local(''),
-       url('/fonts/nunito-v20-latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
-       url('/fonts/nunito-v20-latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
+       url('/fonts/nunito-v20-latin-regular.woff2') format('woff2'),
+       url('/fonts/nunito-v20-latin-regular.woff') format('woff');
 }
 
 /* nunito-700 - latin */
@@ -15,8 +15,53 @@ export const fontFaces = `
   font-style: normal;
   font-weight: 700;
   src: local(''),
-       url('/fonts/nunito-v20-latin-700.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
-       url('/fonts/nunito-v20-latin-700.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
+       url('/fonts/nunito-v20-latin-700.woff2') format('woff2'),
+       url('/fonts/nunito-v20-latin-700.woff') format('woff');
+}
+       
+/* OpenDyslexic Regular */
+@font-face {
+  font-family: 'OpenDyslexic';
+  font-style: normal;
+  font-weight: 400;
+  src: local(''),
+       url('/fonts/OpenDyslexic-Regular.woff') format('woff');
+}
+      
+/* OpenDyslexic Bold */
+@font-face {
+  font-family: 'OpenDyslexic';
+  font-style: normal;
+  font-weight: 700;
+  src: local(''),
+       url('/fonts/OpenDyslexic-Bold.woff') format('woff');
+}
+       
+/* OpenDyslexic Italic */
+@font-face {
+  font-family: 'OpenDyslexic';
+  font-style: italic;
+  font-weight: 400;
+  src: local(''),
+       url('/fonts/OpenDyslexic-Italic.woff') format('woff');
+}
+       
+/* OpenDyslexic Bold Italic */
+@font-face {
+  font-family: 'OpenDyslexic';
+  font-style: italic;
+  font-weight: 700;
+  src: local(''),
+       url('/fonts/OpenDyslexic-Bold-Italic.woff') format('woff');
+}
+       
+/* OpenDyslexic Mono */
+@font-face {
+  font-family: 'OpenDyslexic Mono';
+  font-style: normal;
+  font-weight: 400;
+  src: local(''),
+       url('/fonts/OpenDyslexicMono-Regular.otf') format('opentype');
 }
        
 /* source-code-pro-regular - latin */
@@ -25,8 +70,8 @@ export const fontFaces = `
   font-style: normal;
   font-weight: 400;
   src: local(''),
-       url('/fonts/source-code-pro-v18-latin-regular.woff2') format('woff2'), /* Chrome 26+, Opera 23+, Firefox 39+ */
-       url('/fonts/source-code-pro-v18-latin-regular.woff') format('woff'); /* Chrome 6+, Firefox 3.6+, IE 9+, Safari 5.1+ */
+       url('/fonts/source-code-pro-v18-latin-regular.woff2') format('woff2'),
+       url('/fonts/source-code-pro-v18-latin-regular.woff') format('woff');
 }
 `;